+++ strategies/include/siis/analysers/timeframebaranalyser.h
@@ -33,7 +33,7 @@ public:
             o3d::Double timeframe,
             o3d::Double sourceTimeframe,
             o3d::Int32 depth,
-            o3d::Int32 history,
+            o3d::Double history,
             Price::Method priceMethod=Price::PRICE_CLOSE);
 
     virtual ~TimeframeBarAnalyser() override;
